requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
audio-recorder-streamlit==0.0.8
audioread==3.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
exceptiongroup==1.1.1
ffmpeg-python==0.2.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.0
librosa==0.10.0.post2
llvmlite==0.40.0
markdown-it-py==2.2.0
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
outcome==1.2.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
pooch==1.6.0
protobuf==3.20.3
pyarrow==12.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.21.0
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.30.0
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.10.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
streamlit==1.22.0
streamlit-audiorecorder==0.0.2
streamlit-lottie==0.0.3
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
tornado==6.3.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
undetected-chromedriver==3.4.7
urllib3==2.0.2
validators==0.20.0
wavio==0.0.7
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
